# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /provider/payouts:
    get:
      operationId: listProviderPayouts
      tags: [payouts]
      summary: View provider earnings and payout history
      responses:
        "200": { description: Earnings and payouts; transfers are manual }
    post:
      operationId: requestProviderPayoutApi
      tags: [payouts]
      summary: Request a balance-backed provider payout
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, exclusiveMinimum: 0 }
      responses:
        "201": { description: Payout request recorded; no automatic transfer is made }
...
  /admin/payouts/{id}:
    patch:
      operationId: processAdminPayoutApi
      tags: [admin, payouts]
      summary: Record or reject a manual bank transfer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, note]
              properties:
                decision: { type: string, enum: [COMPLETE, REJECT] }
                transferReference: { type: string, maxLength: 120 }
                proofUrl: { type: string, description: HTTPS transfer proof URL }
                note: { type: string, minLength: 5, maxLength: 1000 }
      responses:
        "200": { description: Manual payout status persisted }
...